Changed floats into doubles in the Epics communication code.
git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@3312 6dcd42c9-f523-4c6d-aada-af552506706e

+++ src/include/drv/fm10Gen.c
@@ -760,7 +760,7 @@ filterModuleD2(FILT_MOD *pFilt,     /* Filter module data  */
   fmInput = 0;
 
   /* Load the filter input testpoint to the input value. */
-  pFilt->data[modNum].filterInput = (float)(filterInput + fmInput);
+  pFilt->data[modNum].filterInput = (filterInput + fmInput);
 
   /* Add excitation signal to input value. */
   fmInput += (double)pFilt->data[modNum].exciteInput;
